refactor: move Illustration into SharedComponents
