+++ src/components/EarlyReleaseUnitsHeader/EarlyReleaseUnitsHeader.tsx
@@ -1,9 +1,9 @@
 import { GridArea } from "../Grid";
 import Grid from "../Grid/Grid";
-import Illustration from "../Illustration/Illustration";
 import { Heading } from "../Typography";
 import Typography from "../Typography/Typography";
 
+import Illustration from "@/components/SharedComponents/Illustration";
 import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
 import Flex from "@/components/SharedComponents/Flex";
