+++ src/components/HomePageTabs/CurriculumTab/CurriculumTab.tsx
@@ -6,7 +6,7 @@ import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
 import Flex from "@/components/SharedComponents/Flex";
 import { GridArea } from "@/components/Grid";
 import Grid from "@/components/Grid/Grid";
-import Illustration from "@/components/Illustration/Illustration";
+import Illustration from "@/components/SharedComponents/Illustration";
 import MaxWidth from "@/components/SharedComponents/MaxWidth";
 import { Heading, LI, UL } from "@/components/Typography";
 import Typography from "@/components/Typography/Typography";
