+++ src/components/Posts/UpcomingWebinarWall/UpcomingWebinarWall.tsx
@@ -1,8 +1,8 @@
 import { FC } from "react";
 
 import { Heading, HeadingTag } from "../../Typography";
-import Illustration from "../../Illustration";
 
+import Illustration from "@/components/SharedComponents/Illustration";
 import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
 import Box from "@/components/SharedComponents/Box";
 import Flex from "@/components/SharedComponents/Flex";
