+++ src/components/SharedComponents/Illustration/Illustration.tsx
@@ -1,7 +1,6 @@
 import { FC } from "react";
 
-import { getIllustrationAsset, IllustrationSlug } from "../../image-data";
-
+import { getIllustrationAsset, IllustrationSlug } from "@/image-data";
 import { CMSImageProps } from "@/components/SharedComponents/CMSImage/CMSImage";
 import CMSImage from "@/components/SharedComponents/CMSImage";
 import Box from "@/components/SharedComponents/Box";
