+++ src/components/SharedComponents/Illustration/index.ts
