+++ src/pages/develop-your-curriculum.tsx
@@ -19,7 +19,7 @@ import GridArea from "@/components/Grid/GridArea";
 import { getSeoProps } from "@/browser-lib/seo/getSeoProps";
 import Cover from "@/components/Cover";
 import BrushBorders from "@/components/SpriteSheet/BrushSvgs/BrushBorders";
-import Illustration from "@/components/Illustration";
+import Illustration from "@/components/SharedComponents/Illustration";
 import { getSizes } from "@/components/SharedComponents/CMSImage/getSizes";
 import getPageProps from "@/node-lib/getPageProps";
 import { PortableTextWithDefaults } from "@/components/PortableText";
